+++ modules/ecs_auto_scaling_group/outputs.tf
@@ -1,3 +1,3 @@
 output "autoscaling_group_name" {
-  value = "${aws_cloudformation_stack.asg.outputs["AsgName"]}"
+  value = "${element(concat(aws_autoscaling_group.asg.*.name, list("")), 0)}"
 }