+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 ### Deprecated
 ### Removed
+
+* Removes code for rolling upgrades of ECS AMIs
+
 ### Fixed
 ### Security