+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 * Changes cache_behaviour to ordered_cache_behaviour as cache_behaviour was removed in AWS provider version 2
+* Uses skip variable for ecr lifecycle policy
 
 ### Deprecated
 ### Removed